Rename Error -> ErrorComponent to prevent naming conflicts
