+++ packages/stjude-cloud-theme-react/stories/Error.stories.js
@@ -1,10 +1,10 @@
 import React from "react";
-import Error from "../src/Error";
+import ErrorComponent from "../src/ErrorComponent";
 
 export default {
   title: "Error",
-  component: Error
+  component: ErrorComponent
 };
 
-export const Default = () => <Error />;
-export const NotFound = () => <Error variant="notFound" />;
+export const Default = () => <ErrorComponent />;
+export const NotFound = () => <ErrorComponent variant="notFound" />;
